// webapp/agnor-backend/src/clientes/interface/cliente.interface.ts
export interface ClienteI {
    id?: number;
    nombre?: string;
    departamento?: string;
    edad?: number;
    fecha_registro?: string;
    clasificacion?: string;
    estado?: string;
    apellidos?: string;
    telefono1?: number;
    telefono2?: number;
    correo?: string;
    tipoanuncio?: string;
    lineanegocio?: string;
    tipopropiedad?: string;
    fechaentrada?: string;
    notas?: string;
    semaforo?: string;
    asignadoa?: string;
    tiempoasignado?: string;
    grupo?: string;
    formadepago?: string;
    montocliente?: string;
    rangoprecios?: string;
  }